Mv types to separate file
