+++ client/dom/index.ts
@@ -30,6 +30,7 @@ export * from './table'
 export * from './table2col'
 export * from './toggleButtons.ts'
 export * from './types/boxplot.ts'
+export * from './types/table.ts'
 export * from './types/toggleButtons.ts'
 export * from './violinRenderer'
 export * from './zoom'
